Start running verlet
